Cleanup: move initialization to the builtin plugin
- instead of initializing the built-in rules, rulesets, and
  formatters in the PluginManager, move the initialization
  into the BuiltinPlugin which is the more logical place for them
  to live
